+++ src/async/tasks.ts
@@ -148,7 +148,7 @@ export const TASKS = new Context<Tasks | undefined>();
 /**
  * Check if there are any pending tasks in the current tasks instance.
  *
- * This can be used in conjuction with {@link waitFor} to indicate if there are any pending tasks.
+ * This can be used in conjuction with {@link Tasks.prototype.waitFor `TASKS.current.waitFor`} to indicate if there are any pending tasks.
  *
  * This is meant to be used for preventing concurrent user interaction in a specific context.
  *
@@ -164,7 +164,7 @@ export function isSelfPending(): boolean {
 /**
  * Check if there are any {@link Tasks.prototype.pending pending} tasks in the current tasks instance or any of it's parents.
  *
- * This can be used in conjunction with {@link waitFor} to disable inputs and buttons while there are any pending tasks.
+ * This can be used in conjunction with {@link Tasks.prototype.waitFor `TASKS.current.waitFor`} to disable inputs and buttons while there are any pending tasks.
  *
  * This is meant to be used for preventing concurrent user interaction in a specific context.
  *
@@ -178,34 +178,14 @@ export function isPending(): boolean {
 }
 
 /**
- * Pretend, that there is a {@link Tasks.prototype.setPending pending} task in the current tasks instance until the current lifecycle is disposed.
- *
- * This is meant to be used for preventing concurrent user interaction in a specific context.
- *
- * @example
- * ```tsx
- * import { TASKS, Tasks, capture, setPending, isPending } from "rvx";
- *
- * TASKS.inject(new Tasks(), () => {
- *   isPending(); // => false
- *   const stop = capture(setPending);
- *   isPending(); // => true
- *   stop();
- *   isPending(); // => false
- * });
- * ```
+ * @deprecated Use `TASKS.current.setPending()` instead. This silently fails if there is no `Tasks` instance in the current context.
  */
 export function setPending(): void {
 	TASKS.current?.setPending();
 }
 
 /**
- * Use the current tasks instance to {@link Tasks.prototype.waitFor wait for} an async function or promise.
- *
- * This is meant to be used for preventing concurrent user interaction in a specific context.
- *
- * @param source The async function or promise to wait for.
- * + If this is a function, it runs {@link isolate isolated}.
+ * @deprecated Use {@link Tasks.prototype.waitFor `TASKS.current.waitFor`} insead. This silently fails if there is no `Tasks` instance in the current context.
  */
 export function waitFor(source: TaskSource): void {
 	TASKS.current?.waitFor(source);