+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 22.4
 + Add `Content` type alias.
 + Add `assertViewState` test utility.
++ `viewNodes` now supports updating already yielded nodes during iteration.
 
 ## 22.3
 + Fix: Errors thrown while notifying observers in a `batch` no longer cause batch dispatching to abort and potentially leave other signal/observer relations in a broken state.